Refactor Entities + add mySQL project with DataGrip

+++ QuickByte/src/main/java/domain/Category.java
@@ -1,4 +1,4 @@
 package domain;
 
-public record Category(Long categoryID, String name, String description) {
+public record Category(int categoryID, String name, String description) {
 }
